Adjust layout of the back of the card
This is an improved and simplified version which solves the issue
of the small font size for the URL. All text has been shifted upwards,
which gives a larger space to the URL. Each text now uses the strong
directional line created by an alignment of triangles as baseline.

I relaxed the constraint giving an equal width to both phone numbers, and I
moved the French number above the US one because of its larger width.
In this new layout, this actually preserves the higher importance of the US
phone number in the hierarchy, which is aligned with a more important
set of triangles than the French numbers.

Also, I increased the padding on the right of the text, which gives more
breathing space and extra safety against irregular cuts of the card.