Define reusable Sierpinski shape in front of the card as well
On the front of the card, the shape appears only once,
vs twice on the back of the card. Still, I extracted
the definition to make both files more similar.

I may then extract common definitions and styles
to separate files to avoid duplicate definitions
in both files and to make changes in a single place.