Extract definition of Sierpiński triangle shape to an external SVG file
An alternative would have been to define sierpinski.svg as a copy
of card-front.svg, then to reference the whole SVG image in
both card-front.svg and card-back.svg as:

  <image x="..." y="..." width="6800" height="10800"
    xlink:href="sierpinski/sierpinski.svg"
  />

Note that the width and height attributes are required on this element.
Also, only the CSS styles defined directly in the referenced SVG file
are taken into account, because based on the browser security policy,
external style sheets are not loaded for referenced images.

This makes it harder to manage CSS styles for the Sierpiński triangle
and make sure that they agree with those of the other shapes such as
the text in the host document.

I preferred to keep all the CSS styles in a single style sheet,
which lead me to favor a reference through the <use> element
rather than an <image>.