Inline all CSS styles and SVG defs/symbols to print to PDF correctly
The new PDF files are now complete, as expected.

The SVG files are now very redundant. I kept some of that redundancy
visible, by keeping separate <style> and <defs> elements in test-sheet.svg
for the parts previously imported in layers from sierpinski/sierpinski.svg
and sierpinski/styles.css on one hand, and from card-front.svg and
card-back.svg on the other hand.

I have removed the shared definitions in the sierpinski folder,
which are now unused. I am sad to see it go and I hope that this
approach will become viable in the future.

The advantage is that each SVG file is now self-sufficient and
independent, and in a more widely supported format.