Rename back to front and vice versa
For consistency with the rest of the project.
I forgot that I had previously designed the front as the part without text,
the top of the Sierpiński triangle, while the part with the text is on the
back of the card, with the bottom of the Sierpiński triangle.