Position phone numbers diagonally
The US phone number comes first, in a descending diagonal.
The French phone number comes last, in an ascending diagonal.

Phone numbers are rotated by an angle of 38.45 degrees,
computed as the inverse tangent of the ratio of the card sizes:

  tan(α) = 5400/6800

  i.e. α = arctan(27/34)
         ~ 38.45°

The rotation center is the middle of the '+' sign at the beginning
of each phone number, a reference point which is visually explicit,
as a cross often marks a point on a mathematical diagram.

The size of the text for the country codes, US and FR, was reduced,
to de-emphasize them relatively to the digits of the phone number.
I first tried to apply small caps, but they it made no difference
with the select font, Gill Sans. The aim of the reduced font size
is to reach a similar effect. The font weight is reduced slightly
as well, but the difference is not visually disturbing in this case,
since these are the only alphabetical letters, as opposed to digits
and punctuation signs. It could be compensated by increasing the font
weight for the country codes.

The parentheses and the dashes in the French number were coming as
too heavy in this oblique orientation. I switched them to the serif
font "Johanna Nova", designed by Eric Gill like the serif font Gill
Sans. These two fonts work well together, and it so happens that the
punctuation signs are lighter in the serif version. I have purchased
the "Johanna Nova" font, here in medium weight, it is not available
as a system font on macOS like Gill Sans. I have configured a fallback
font of 'serif' which also works well, to allow a similar effect when
the font is not available.

I replaced space characters in the phone numbers with non-breaking
narrow versions. This makes the phone numbers more compact.

As illustrated below, US phone number is one character shorter
than the French phone number:

 +1 (272) 999-1984 (US)
+33 (0)644-643-642 (FR)

In order to compensate that difference, I used different kinds
of dashes in the two phone numbers:

  * an en dash in the US phone number
  * non-breaking hyphens in the French phone number

The en dash is usually the marker of a range of values. This is
not incompatible with the form at the end of this specific US phone
numbers, which does look like a range from 999 to 1984.

Similarly, non-breaking hyphens indicate closely related words
that must not be separated by line breaks, which is also relevant
for the three chunks separated by hyphens in the French phone number.