Add role name, Instructional Designer, on card front
Each word follows a diagonal orientation similar
to the phone numbers on the back of the card.

The word Instructional, on top, follows an ascending diagonal.
The word Designer, below, follows a descending diagonal.

The initial letter of each word is rotated back to the upright
orientation. The two initial letters for the abbreviation ID,
which stands commonly for Instructional Designer.

Most of the text is set in sans-serif, i.e. Gill Sans,
while the initials and a few other letters are set in
the serif typeface Johanna Nova.

These extra letters, C, I in the top word and E, R
in the bottom word, are the four letters of my given
name, ERIC.