Apply bleed of ⅛in or 400 units on each side of the card
I have noticed that the mirror transformation applied to the
Sierpiński triangle is shifted to the right. This reflects the
fact the 'center' in the transform-origin relates to a partial
viewBox starting at the corner 0,0 rather than the whole viewBox
which starts at an offset of -400,-400 to avoid updating all
the coordinates of the drawing.

I will replace this keyword with the coordinates of the central
point, disregarding the added bleed of 400 units on each side:

  * left = 6800 / 2 = 3400
  * top = 10800 / 2 = 5400