Add comment which computes the viewBox scaling factor
The key insight is that the scaling factor is computed
by considering that the width/height of the document
is in pixels, converting inches using 1 in = 96 px,
and that the user units are defining the new pixels.

The scaling factor is then computed, for each of the
document dimensions, width and height, as the ratio
of the number of pixels by the number of user units.

References:

[1] 2018-04-24 answer by ccprog
to SVG styles with absolute units
on Stack Overflow
https://stackoverflow.com/a/50010403

[2] 2017-01-29 answer by FWrnr
to SVG and viewBox values
on Graphic Design Stack Exchange
https://graphicdesign.stackexchange.com/a/84170

[3] 5.2. Absolute Lengths: the cm, mm, Q, in, pt, pc, px units
in CSS Values and Units Module Level 3
https://www.w3.org/TR/css3-values/#absolute-lengths

[4] 8.6. The ‘viewBox’ attribute
in Chapter 8: Coordinate Systems, Transformations and Units
in Scalable Vector Graphics (SVG) 2
https://www.w3.org/TR/SVG/coords.html#ViewBoxAttribute

[5] 8.2. Computing the equivalent transform of an SVG viewport
in Chapter 8: Coordinate Systems, Transformations and Units
in Scalable Vector Graphics (SVG) 2
https://www.w3.org/TR/SVG/coords.html#ComputingAViewportsTransform