Reduce stroke width from 0.7pt to 0.5pt to get a light gray effect
I was planning to change the color of the lines to a gray,
but the same effect was achieved by reducing the stroke width.

I found this interesting reference which lead me to this choice,
explaining how anything less than 100% black is produced using
halftone dots which will reduce the accuracy of the lines:

  (…)
  the offset print uses 4 colors: cyan magenta yellow and black.
  lets say your line would be 50% black, also known as gray.
  the printer would be still using only black ink to produce the
  gray line, but instead of printing it with solid color (black)
  the illusion of gray will be produced through a halftone pattern…
  (…)
  if you want to use thin lines dont go bellow 0.25pt.
  and thats if your lines are solid, one component CMYK color.
  For tints and mixed colors I wouldn’t go bellow 0.4pt.
  — matti1, December 10th, 2011
  https://forums.envato.com/t/whats-the-minimum-line-width-for-print/75593/3

Having only a single color, 100% black, will simplify the design.
The conversion from RGB black to CMYK black will be straightforward.