Clip drawings to document viewBox dimensions
Otherwise, Inkscape displays and exports drawings
outside the viewBox of the document.

In order to clip to the document dimensions, I created
a rectangle, which happened to be filled in red, from
the top-left corner intersection of the guides to the
bottom-right corner intersection of the guides around
the document viewBox.

I renamed this rectangle to "clipping", to clarify, then
I selected it and the layer with all the drawings below.
In the menu, I then selected Object > Clip > Set. This
made the "clipping" object disappear and clipped the layer
of drawings to the viewBox area expected to be the sole
visible in these documents.