Replace clipping with visible frame for the document viewBox
The problem is not the export from Inkscape but the import in Scribus.
Some vector shapes are drawn outside of the document viewBox, which is
to be expected. But Scribus imports the whole shapes, without taking
into account the boundaries of the document viewBox, which makes a mess.

Having a visible frame will allow me to size and position the imported
document correctly in Scribus, and to crop the contents to the given area.