Import EPS for each side of the card in Scribus
On each page, File > Import > Get Vector File…
then select the EPS file exported from Scribus
and click anywhere in the page.

The imported file extends beyond the boundaries
of the document viewBox area. We will fix this
in the next step.

At this step, I just repositioned the imported
vector images towards the center of the page,
very approximately.