Revert "Import EPS for each side of the card in Scribus"
There is no easy way to give the correct size accurately
to the imported shapes unless all the shapes are exported.

By exporting the whole drawing to EPS, I will be able to get
the size of the whole, in inches, in Inkscape, and scale the
imported shapes to the same size, in inches, in Scribus.