Resize and position front of the card
The width and height are given by selecting all the shapes
in Inkscape. Setting one with values linked sets the other
to the expected value.

The x position corresponds to the left edge of the bleed,
i.e. the negative value -0.125in in Scribus.

The y offset corresponds to the whole height minus
the height of the card without bleed, divided by two:
(6.750in - 3.375in) / 2 = 1.6875in

The y position is that offset in negative: -1.6875in.

These offsets and dimensions bring the yellow area
in agreement with the surface of the page with bleeds.