Resize and position back of the card
The width and height are given by selecting all the shapes
in Inkscape. Setting one value while preserving ratio sets
the other to the expected value.

The x offset is the full width minus the width of the card
with a single bleed, nothing extending past the right bleed,
which is empty.

This gives a negative x position of: -2.125in.
The y offset is identical to the one computed for the front
of the card: -1.6875in.