Extract yellow frames and move them to background layer
I selected each front and back layer in turn.
On each layer, I select the group of imported shapes,
ungrouped, and with the shift key pressed, I unselected
carefully just the yellow frame, by clicking in one of
the empty bleed areas where it is left alone. I then
regrouped the rest of the shapes without the yellow frame.

I then cut and pasted the yellow frames onto the separate
Background layer, which I have configure so that it will not
be printed.