Create clipping shapes for the front and back of the card
On two new layers, one for each of the clipping shapes,
I have drawn a large rectangle, filled with cyan, which
large enough to cover all the imported shapes for the
side of the card below, and a bit more on each side.

I made a copy of the yellow background frame for that
side of the card on top of the cyan rectangle. Then
I applied the subtraction path operation to create
a clipping mask from the cyan rectangle with a whole:

1. select the cyan shape
2. select the yellow shape on top
3. go to Item > Path Tools… > Path Operations…
4. select subtraction, the second operation