Clip each side of the card to the expected viewBox area
On each layer, make a copy of the corresponding clipping shape
and push it to the bottom level, below all imported shapes.

Select the clipping shape first, then with shift pressed, select
the group of imported shapes. Apply the clipping using the subtraction
operation in Item > Path Tools… > Path Operations…

I also moved the layers with the copy of the clipping shapes
below the layer with the copy of all original unclipped shapes.