Move unclipped shapes of the back of the card to a separate layer
I noticed a problem with the width of the strokes of the drawings.
They seem to be printed as hairline, as thin as possible at any
zoom level, which does not result in visible lines on my printer.

The stroke width should be set to 0.5pt instead. I expect that
ungrouping the clipped groups will result in the loss of clipping,
which is why I prefer to have a copy of the unclipped shapes on
hand instead.