Trim the triangle shapes for each side of the card, on new layers
I proceeded in five steps:

1. I made a copy of all the triangles shapes on each side of the card
2. I deleted triangles falling entirely outside of the page and bleed area
3. I grouped all the remaining triangles
4. I copied the trimming shapes, which I lowered to the lowest level
5. I select the triangles group first, then, with shift pressed,
   the trimming shape, and I applied the subtraction path operation.

As a result, triangles have now been trimmed for printing with bleed
on each side of the card. I preferred to remove extra shapes, falling
entirely outside of the bleed area, to avoid the risk of having them
carried hidden, out of the page, into the exported PDF.

I have also deleted the previous version of the trimmed sides of the
card, which contained both the text and the shapes grouped together.