Import updated text on front of the card
I imported the EPS file, then resized it to match the position
and dimensions of the yellow frame alone in the last layer.

After resizing and positioning, I ungrouped the imported shapes
and deleted the yellow frame, leaving only the updated text.

I did all this in a new layer, while keeping the previous version
of the text on the front of the card in another layer. After I
completed and checked the position of the updated text with regards
to the position of the previous one, I deleted the layer with the
old version of the text.